backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.7.0
cohere==4.34
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-frontmatter==1.0.0
psutil==5.9.5